fix(codegraph): auto-init secondary project paths
