+++ packages/codegraph-mcp/src/proxy.ts
@@ -147,7 +147,7 @@ async function forwardClientToCodegraph(options: ClientForwardOptions, state: Pr
     const projectPath = projectPathFromToolCall(message.payload)
     if (projectPath !== null) {
       try {
-        const refreshed = await options.synchronizer.refresh(projectPath, false)
+        const refreshed = await options.synchronizer.refresh(projectPath, options.autoInit)
         if (!refreshed) {
           const parsed = isPlainRecord(message.payload) ? message.payload : {}
           writeStdioJsonRpcResponse(
